add flags to compile 2 counters
