+++ examples/upgradable-examples/upgradable_base/Cargo.toml
@@ -18,3 +18,7 @@ workspaces = "0.6"
 tokio = { version = "1.1", features = ["rt", "macros"] }
 serde_json = "1.0.74"
 near-plugins-test-utils = { path = "../../near-plugins-test-utils" }
+
+[features]
+counter1 = []
+counter2 = []
