--- examples/upgradable-examples/upgradable_base_second/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "upgradable_base_second"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-near-sdk = "4.0.0"
-near-plugins = { path = "../../../near-plugins" }
-near-plugins-derive = { path = "../../../near-plugins-derive" }
-borsh = "0.9.3"
-
-[dev-dependencies]
-near-primitives = "0.14.0"
-workspaces = "0.6"
-tokio = { version = "1.1", features = ["rt", "macros"] }
-serde_json = "1.0.74"
-
-
