--- examples/upgradable-examples/upgradable_base_second/build.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-RUSTFLAGS='-C link-arg=-s' cargo build --target wasm32-unknown-unknown --release
